+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GameService.java
@@ -159,6 +159,8 @@ public class GameService implements InitializingBean, ApplicationListener<Applic
     emulatorService.loadEmulators();
     mameRomAliasService.clearCache();
     highscoreService.refreshAvailableScores();
+    gameCachingService.clearCache();
+    getKnownGames(-1);
     return true;
   }